webhook/pyproject.toml
[project]
name = "webhook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "line-bot-sdk>=3.21.0",
    "shared",
    "sqlalchemy>=2.0.45",
]

[tool.uv.sources]
shared = { workspace = true }
